+++ app/src/renderer/vuex/modules/distribution.js
@@ -61,6 +61,14 @@ export default ({ node }) => {
     }
   }
   const actions = {
+    async reconnected({ rootState, state, dispatch }) {
+      if (state.loading && rootState.session.signedIn) {
+        await dispatch(`getTotalRewards`)
+      }
+    },
+    async initializeWallet({ dispatch }) {
+      await dispatch(`getTotalRewards`)
+    },
     resetSessionData({ rootState }) {
       rootState.distribution = JSON.parse(JSON.stringify(emptyState))
     },
@@ -73,12 +81,12 @@ export default ({ node }) => {
         const rewards = coinsToObject(rewardsArray)
         commit(`setTotalRewards`, rewards)
         commit(`setDistributionError`, null)
-        state.loaded = true
       } catch (error) {
         Sentry.captureException(error)
         commit(`setDistributionError`, error)
       }
       state.loading = false
+      state.loaded = true
     },
     async withdrawAllRewards(
       { rootState: { wallet }, dispatch },
@@ -92,6 +100,14 @@ export default ({ node }) => {
       })
       await dispatch(`getTotalRewards`)
     },
+    async getRewardsFromAllValidators({ state, dispatch }, validators) {
+      state.loading = true
+      await Promise.all(validators.map(validator =>
+        dispatch(`getRewardsFromValidator`, validator.operator_address)
+      ))
+      state.loading = false
+      state.loaded = true
+    },
     async getRewardsFromValidator({ state, rootState: { session }, commit }, validatorAddr) {
       state.loading = true
       try {
@@ -99,12 +115,12 @@ export default ({ node }) => {
         const rewards = coinsToObject(rewardsArray)
         commit(`setDelegationRewards`, { validatorAddr, rewards })
         commit(`setDistributionError`, null)
-        state.loaded = true
       } catch (error) {
         Sentry.captureException(error)
         commit(`setDistributionError`, error)
       }
       state.loading = false
+      state.loaded = true
     },
     async withdrawRewardsFromValidator(
       { rootState: { wallet }, commit, dispatch },
@@ -130,12 +146,12 @@ export default ({ node }) => {
         const parameters = await node.getDistributionParameters()
         commit(`setDistributionParameters`, parameters)
         commit(`setDistributionError`, null)
-        state.loaded = true
       } catch (error) {
         Sentry.captureException(error)
         commit(`setDistributionError`, error)
       }
       state.loading = false
+      state.loaded = true
     },
     async getOutstandingRewards({ commit }) {
       state.loading = true
@@ -144,12 +160,12 @@ export default ({ node }) => {
         const oustandingRewards = coinsToObject(oustandingRewardsArray)
         commit(`setOutstandingRewards`, oustandingRewards)
         commit(`setDistributionError`, null)
-        state.loaded = true
       } catch (error) {
         Sentry.captureException(error)
         commit(`setDistributionError`, error)
       }
       state.loading = false
+      state.loaded = true
     }
   }